backend/requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# ── Pydantic / settings ───────────────────────────────────────────────────────
pydantic[email]==2.10.3
pydantic-settings==2.6.1

# ── Security / auth ───────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==43.0.3

# ── Object storage ────────────────────────────────────────────────────────────
boto3==1.35.82

# ── Task queue ────────────────────────────────────────────────────────────────
celery[redis]==5.4.0
redis==5.2.1

# ── AI / OpenAI ───────────────────────────────────────────────────────────────
openai==1.57.2

# ── Video processing ──────────────────────────────────────────────────────────
ffmpeg-python==0.2.0

# ── HTTP client ───────────────────────────────────────────────────────────────
httpx==0.28.1

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
tenacity==9.0.0
